added class page

+++ src/EepromHelper.h
@@ -11,6 +11,6 @@ void write16b( int value, int addr = 0);
 int read16b(int addr);
 
 void writeEprom(int temp, int speed);
-void readEprom(double &temp, int &speed);
+void readEprom(double &temp, float &speed);
 
 #endif
